compile python files in /usr/lib/ckan/default/src
The www-data user will not have permissions to write to the virtualenv
directory. Forcing compilation of files as our packages are arch
specific anyway.

We'll get a slightly faster startup. If the pyc files ever do get
created and were not in the package they will not be removed if the
package is removed.

Additionally we will also avoid any random SELinux denial messages.