Switch to building packages using a Docker image
This replaces the old Ansible / Vagrant setup with Docker. All the
building steps are run in the relevant `ubuntu:X.Y` image, and the
resulting deb package is copied to the client host. As with the old
scripts, you can pass build args to control the process. The full
command is:

```
docker buildx build \
    --output type=local,dest=.  \
    --build-arg CKAN_VERSION=2.11 \
    --build-arg CKAN_BRANCH=dev-v2.11 \
    --build-arg DATAPUSHER_VERSION=0.0.21 \
    --build-arg DISTRIBUTION=noble \
    --build-arg UBUNTU_VERSION=24.04 \
    .
```

For now we are sticking with a single Dockerfile for all Ubuntu versions
as there is only one minor difference in focal, but if there are more
differences added in the future we will switch to different Dockerfiles.